Semestr_5/JavaProjects/bookstore/src/main/java/com/example/bookstore/service/AuthorService.java
package com.example.bookstore.service;

import com.example.bookstore.model.Author;
import com.example.bookstore.repository.AuthorRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class AuthorService {

    @Autowired
    private AuthorRepository authorRepository;

    public List<Author> getAllAuthors() {
        return authorRepository.findAll();
    }

    public Optional<Author> getAuthorById(Long id) {
        return authorRepository.findById(id);
    }

    public Author saveAuthor(Author author) {
        // Перевірка на наявність автора з таким самим ім'ям та прізвищем
        if (authorRepository.existsByNameAndSurname(author.getName(), author.getSurname())) {
            throw new IllegalArgumentException("Taki autor już istnieje!");
        }
        return authorRepository.save(author);
    }

    public void deleteAuthor(Long id) {
        authorRepository.deleteById(id);
    }
}
